Rename ManageTilesView to ManageTiles and move to ui package
